Replace WebtoonReader with latest version
